AOP 적용 및 ArgumentResolver를 이용한 리팩토링

+++ src/main/java/sangwon/wead/controller/CommentController.java
@@ -1,37 +1,41 @@
 package sangwon.wead.controller;
 
 
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.SessionAttribute;
+import sangwon.wead.argument.annotation.Referer;
+import sangwon.wead.argument.annotation.UserId;
+import sangwon.wead.aspect.annotation.CheckLogin;
+import sangwon.wead.aspect.annotation.NeedLogin;
+import sangwon.wead.exception.ClientFaultException;
+import sangwon.wead.exception.NonexistentPostException;
 import sangwon.wead.service.DTO.CommentUploadForm;
 import sangwon.wead.service.CommentService;
-
-import static sangwon.wead.controller.util.AlertPageRedirector.redirectAlertPage;
+import sangwon.wead.service.PostService;
+import sangwon.wead.util.AlertPageRedirector;
 
 @Controller
 @RequiredArgsConstructor
 public class CommentController {
 
     private final CommentService commentService;
+    private final PostService postService;
 
-
+    @NeedLogin
     @PostMapping("/comment/upload/{postId}")
-    public String upload(HttpServletRequest request,
-                         @SessionAttribute(name = "userId", required = false) String userId,
-                         @PathVariable("postId") Long postId,
+    public String upload(@PathVariable("postId") Long postId,
                          @RequestParam("content") String content,
+                         @UserId String userId,
+                         @Referer String referer,
                          Model model) {
-        // 이전 URL
-        String referer = request.getHeader("referer");
-
-        // 로그인을 하지 않은 경우
-        if(userId == null) return redirectAlertPage("로그인을 먼저 해주세요.", referer, model);
+        // 게시글이 존재하지 않을 경우
+        if(!postService.checkPostExistence(postId)) throw new NonexistentPostException();
+        // 댓글 내용이 비었을 경우
+        if(content.isBlank()) return AlertPageRedirector.redirectAlertPage("댓글을 작성해주세요.", referer, model);
 
         // 댓글 업로드
         CommentUploadForm commentUploadForm = CommentUploadForm.builder()
@@ -43,11 +47,15 @@ public class CommentController {
         return "redirect:" + referer;
     }
 
+    @CheckLogin
     @PostMapping("/comment/delete/{commentId}")
-    public String delete(HttpServletRequest request,
-                         @PathVariable("commentId") Long commentId) {
-        // 이전 URL
-        String referer = request.getHeader("referer");
+    public String delete(@PathVariable("commentId") Long commentId,
+                         @UserId String userId,
+                         @Referer String referer) {
+        // 존재하지 않는 댓글일 경우
+        if(!commentService.checkCommentExistence(commentId)) throw new ClientFaultException();
+        // 삭제 권한이 없을 경우
+        if(!commentService.getCommentInfo(commentId).getUserId().equals(userId)) throw new ClientFaultException();
 
         // 댓글 삭제
         commentService.deleteComment(commentId);
