+++ src/main/java/sangwon/wead/controller/PostController.java
@@ -1,6 +1,5 @@
 package sangwon.wead.controller;
 
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -9,15 +8,22 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import sangwon.wead.argument.annotation.Referer;
+import sangwon.wead.argument.annotation.UserId;
+import sangwon.wead.aspect.annotation.CheckLogin;
+import sangwon.wead.aspect.annotation.NeedLogin;
 import sangwon.wead.controller.DTO.*;
+import sangwon.wead.exception.ClientFaultException;
+import sangwon.wead.exception.NonexistentPostException;
 import sangwon.wead.service.DTO.*;
 import sangwon.wead.service.PostService;
 import sangwon.wead.service.CommentService;
 import sangwon.wead.service.book.BookService;
+import sangwon.wead.util.AlertPageRedirector;
 
 import java.util.List;
 
-import static sangwon.wead.controller.util.AlertPageRedirector.redirectAlertPage;
+import static sangwon.wead.util.AlertPageRedirector.redirectAlertPage;
 
 
 @Controller
@@ -32,13 +38,14 @@ public class PostController {
     @GetMapping( "/post")
     public String postList(@RequestParam(value = "page", required = false, defaultValue = "1") int pageNumber,
                        Model model) {
-
+        // 페이지가 범위를 벗어날 경우 재조정
         if(pageNumber < 1) pageNumber = 1;
         Page<PostInfo> page = postService.getPostInfoPage(pageNumber-1, 10);
         int totalPages = page.getTotalPages();
         if(totalPages == 0) totalPages = 1;
         if(pageNumber > totalPages) page = postService.getPostInfoPage(totalPages-1, 10);
 
+        // 게시글 리스트
         model.addAttribute("postList", page.getContent().stream().map(postInfo -> new PostLine(postInfo)));
         model.addAttribute("pageBar", new PageBar(page, 10));
         return "page/post-list";
@@ -46,9 +53,11 @@ public class PostController {
 
 
     @GetMapping("/post/{postId}")
-    public String post(@SessionAttribute(name = "userId", required = false) String userId,
+    public String post(@UserId(required = false) String userId,
                        @PathVariable("postId") Long postId,
                        Model model) {
+        // 게시글이 존재하지 않을 경우
+        if(!postService.checkPostExistence(postId)) throw new NonexistentPostException();
 
         // 조회수 올리기
         postService.increasePostViews(postId);
@@ -65,50 +74,44 @@ public class PostController {
         return "page/post";
     }
 
-
+    @NeedLogin
     @GetMapping("/post/upload")
-    public String uploadForm(HttpServletRequest request,
-                             @SessionAttribute(name = "userId", required = false) String userId,
-                             @RequestParam("isbn") String isbn,
-                             Model model) {
-
-        // 이전 URL
-        String referer = request.getHeader("referer");
-
-        // 로그인을 하지 않은 경우
-        if(userId == null) return redirectAlertPage("로그인을 먼저 해주세요.", referer, model);
-
+    public String uploadForm(@RequestParam("isbn") String isbn, Model model) {
+        // 수정 폼 전송
         BookInfo bookInfo = bookService.getBookInfo(isbn);
         model.addAttribute("bookTitle", bookInfo.getTitle());
         model.addAttribute("isbn", bookInfo.getIsbn());
         return "page/post-upload";
     }
 
-
+    @CheckLogin
     @PostMapping("/post/upload")
-    public String upload(HttpServletRequest request,
-                         @SessionAttribute(name = "userId", required = false) String userId,
-                         @Valid @ModelAttribute PostUploadParam postUploadParam,
+    public String upload(@Valid @ModelAttribute PostUploadParam postUploadParam,
                          BindingResult bindingResult,
+                         @UserId String userId,
+                         @Referer String referer,
                          Model model,
                          RedirectAttributes redirectAttributes) {
-
-        // 이전 URL
-        String referer = request.getHeader("referer");
-
         // 제목 및 내용이 입력되지 않을 경우
         if(bindingResult.hasErrors()) return redirectAlertPage("제목 및 내용을 모두 입력해주세요.", referer, model);
 
+        // 게시글 업로드
         Long postId = postService.uploadPost(postUploadParam.toPostUploadForm(userId));
         redirectAttributes.addAttribute("postId", postId);
         return "redirect:/post/{postId}";
     }
 
-
+    @CheckLogin
     @GetMapping("/post/update/{postId}")
     public String updateForm(@PathVariable("postId") Long postId,
+                             @UserId String userId,
                              Model model) {
+        // 게시글이 존재하지 않을 경우
+        if(!postService.checkPostExistence(postId)) throw new ClientFaultException();
+        // 수정 권한이 없을 경우
+        if(!postService.getPostInfo(postId).getUserId().equals(userId)) throw new ClientFaultException();
 
+        // 수정 폼 전송
         PostInfo postInfo = postService.getPostInfo(postId);
         BookInfo bookInfo = bookService.getBookInfo(postInfo.getIsbn());
 
@@ -119,28 +122,41 @@ public class PostController {
     }
 
 
+    @CheckLogin
     @PostMapping("/post/update/{postId}")
-    public String update(HttpServletRequest request,
-                         @PathVariable("postId") Long postId,
+    public String update(@PathVariable("postId") Long postId,
                          @Valid @ModelAttribute PostUpdateParam postUpdatePram,
                          BindingResult bindingResult,
-                         RedirectAttributes redirectAttributes,
-                         Model model) {
-
-        // 이전 URL
-        String referer = request.getHeader("referer");
-
+                         @UserId String userId,
+                         @Referer String referer,
+                         Model model,
+                         RedirectAttributes redirectAttributes) {
+        // 게시글이 존재하지 않을 경우
+        if(!postService.checkPostExistence(postId)) throw new ClientFaultException();
+        // 수정 권한이 없을 경우
+        if(!postService.getPostInfo(postId).getUserId().equals(userId)) throw new ClientFaultException();
         // 제목 및 내용이 입력되지 않을 경우
-        if(bindingResult.hasErrors()) return redirectAlertPage("제목 및 내용을 모두 입력해주세요.", referer, model);
+        if(bindingResult.hasErrors()) {
+            return AlertPageRedirector.redirectAlertPage("제목과 내용을 모두 입력해주세요.", referer, model);
+        }
 
+        // 게시글 수정
         PostInfo postInfo = postService.getPostInfo(postId);
         postService.updatePost(postUpdatePram.toPostUpdateForm(postInfo.getPostId()));
         redirectAttributes.addAttribute("postId", postId);
         return "redirect:/post/{postId}";
     }
 
+    @CheckLogin
     @PostMapping("/post/delete/{postId}")
-    public String delete(@PathVariable("postId") Long postId) {
+    public String delete(@PathVariable("postId") Long postId,
+                         @UserId String userId) {
+        // 게시글이 존재하지 않을 경우
+        if(!postService.checkPostExistence(postId)) throw new ClientFaultException();
+        // 삭제 권한이 없을 경우
+        if(!postService.getPostInfo(postId).getUserId().equals(userId)) throw new ClientFaultException();
+
+        // 게시글 삭제
         postService.deletePost(postId);
         return "redirect:/post";
     }
